==> analysis/03-feature-selection.Rmd <==
---
title: "Feature selection"
author: "James Ashmore"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output: workflowr::wflow_html
editor_options:
  chunk_output_type: console
---

## Introduction

In this document we are going to select features for downstream processing.

## Setup

Set chunk options:

```{r knitr}
knitr::opts_chunk$set(
  autodep    = TRUE,
  cache      = TRUE,
  cache.path = "cache/03-feature-selection.Rmd/",
  dev        = "png",
  error      = FALSE,
  message    = FALSE,
  warning    = FALSE
)
```

Load required packages:

```{r pacman}
pacman::p_load(
  BiocSingular,
  org.Mm.eg.db,
  biomaRt,
  eulerr,
  readr,
  here,
  pheatmap,
  scales,
  scater,
  scran
)
```

Read object from file:

```{r}
rds <- here("data/02-normalization.rds")

sce <- read_rds(rds)
```

## Whitelist

Create a whitelist of genes with which to perform feature selection:

```{r}
rowData(sce)$whitelist <- TRUE
```

Remove female sex genes from whitelist:

```{r}
idx <- rowData(sce)$gene_id == "ENSMUSG00000086503" # Xist

rowData(sce)$whitelist[idx] <- FALSE
```

Remove male sex genes from whitelist:

```{r}
mrt <- useMart("ensembl")

mrt <- useDataset("mmusculus_gene_ensembl", mart = mrt)

out <- getBM(c("ensembl_gene_id", "chromosome_name"), mart = mrt)

out <- subset(out, chromosome_name == "Y")

idx <- rowData(sce)$gene_id %in% out$ensembl_gene_id

rowData(sce)$whitelist[idx] <- FALSE
```

Remove cell-cycle genes from whitelist:

```{r}
org <- select(org.Mm.eg.db, keys = "GO:0007049", keytype = "GOALL", column = "ENSEMBL")

idx <- rowData(sce)$gene_id %in% org$ENSEMBL

rowData(sce)$whitelist[idx] <- FALSE
```

## Measuring variation {.tabset}

### modelGeneVar

Model the per-gene variance

```{r}
fit <- modelGeneVar(sce)

dat <- as.data.frame(fit)
  
ggplot(dat, aes(mean, total)) +
  geom_point(colour = "grey") +
  geom_line(aes(y = tech)) +
  labs(x = "Mean of log-expression", y = "Variance of log-expression") +
  theme_minimal()

metadata(sce)$modelGeneVar <- fit
```

### modelGeneCV2

Model the per-gene CV2:

```{r}
fit <- modelGeneCV2(sce)

dat <- as.data.frame(fit)
  
ggplot(dat, aes(mean, total)) +
  geom_point(colour = "grey") +
  geom_line(aes(y = trend)) + 
  scale_x_log10() + 
  scale_y_log10() + 
  labs(x = "Mean of log-expression", y = "Variance of log-expression") +
  theme_minimal()

metadata(sce)$modelGeneCV2 <- fit
```

### modelGeneVarWithSpikes

Model the per-gene variance with spike-ins:

```{r}
fit <- modelGeneVarWithSpikes(sce, "ERCC")

alt <- metadata(fit)

df1 <- as.data.frame(fit)

df2 <- data.frame(mean = alt$mean, total = alt$var)
  
ggplot(df1, aes(mean, total)) +
  geom_point(colour = "grey") +
  geom_line(aes(mean, tech)) +
  geom_point(data = df2, aes(mean, total), colour = "red") +
  labs(x = "Mean of log-expression", y = "Variance of log-expression") +
  theme_minimal()

metadata(sce)$modelGeneVarWithSpikes <- fit
```

### modelGeneCV2WithSpikes

Model the per-gene CV2 with spike-ins:

```{r}
fit <- modelGeneCV2WithSpikes(sce, "ERCC")

alt <- metadata(fit)

df1 <- as.data.frame(fit)

df2 <- data.frame(mean = alt$mean, cv2 = alt$cv2)
  
ggplot(df1, aes(mean, total)) +
  geom_point(colour = "grey") +
  geom_line(aes(mean, trend)) +
  geom_point(data = df2, aes(mean, cv2), colour = "red") +
  scale_x_log10() + 
  scale_y_log10() + 
  labs(x = "Mean of log-expression", y = "Variance of log-expression") +
  theme_minimal()

metadata(sce)$modelGeneCV2WithSpikes <- fit
```

### modelGeneVarByPoisson

Model the per-gene variance with Poisson noise:

```{r}
fit <- modelGeneVarByPoisson(sce)

dat <- as.data.frame(fit)

ggplot(dat, aes(mean, total)) +
  geom_point(colour = "grey") +
  geom_line(aes(y = tech)) +
  labs(x = "Mean of log-expression", y = "Variance of log-expression") +
  theme_minimal()

metadata(sce)$modelGeneVarByPoisson <- fit
```

## Selecting genes {.tabset}

### modelGeneVar

Identify highly variable genes, based on modelGeneVar statistics:

```{r}
fit <- metadata(sce)$modelGeneVar

dat <- fit[rowData(sce)$whitelist, ]

hvg <- getTopHVGs(dat, var.field = "bio", n = 2000)

sel <- ifelse(rownames(fit) %in% hvg, TRUE, FALSE)

metadata(sce)$modelGeneVar$hvg <- sel
```

Plot expression of highly variable genes, based on modelGeneVar statistics:

```{r}
plotHeatmap(sce, hvg, center = TRUE, symmetric = TRUE, show_rownames = FALSE, show_colnames = FALSE)
```

### modelGeneCV2

Identify highly variable genes, based on modelGeneCV2 statistics:

```{r}
fit <- metadata(sce)$modelGeneCV2

dat <- fit[rowData(sce)$whitelist, ]

hvg <- getTopHVGs(dat, var.field = "ratio", n = 2000)

sel <- ifelse(rownames(fit) %in% hvg, TRUE, FALSE)

metadata(sce)$modelGeneCV2$hvg <- sel
```

Plot expression of highly variable genes, based on modelGeneCV2 statistics:

```{r}
plotHeatmap(sce, hvg, center = TRUE, symmetric = TRUE, show_rownames = FALSE, show_colnames = FALSE)
```

### modelGeneVarWithSpikes

Identify highly variable genes, based on modelGeneVarWithSpikes statistics:

```{r}
fit <- metadata(sce)$modelGeneVarWithSpikes

dat <- fit[rowData(sce)$whitelist, ]

hvg <- getTopHVGs(dat, var.field = "bio", n = 2000)

sel <- ifelse(rownames(fit) %in% hvg, TRUE, FALSE)

metadata(sce)$modelGeneVarWithSpikes$hvg <- sel
```

Plot expression of highly variable genes, based on modelGeneVarWithSpikes statistics:

```{r}
plotHeatmap(sce, hvg, center = TRUE, symmetric = TRUE, show_rownames = FALSE, show_colnames = FALSE)
```

### modelGeneCV2WithSpikes

Identify highly variable genes, based on modelGeneCV2WithSpikes statistics:

```{r}
fit <- metadata(sce)$modelGeneCV2WithSpikes

dat <- fit[rowData(sce)$whitelist, ]

hvg <- getTopHVGs(dat, var.field = "ratio", n = 2000)

sel <- ifelse(rownames(fit) %in% hvg, TRUE, FALSE)

metadata(sce)$modelGeneCV2WithSpikes$hvg <- sel
```

Plot expression of highly variable genes, based on modelGeneCV2WithSpikes statistics:

```{r}
plotHeatmap(sce, hvg, center = TRUE, symmetric = TRUE, show_rownames = FALSE, show_colnames = FALSE)
```

### modelGeneVarByPoisson

Identify highly variable genes, based on modelGeneVarByPoisson statistics:

```{r}
fit <- metadata(sce)$modelGeneVarByPoisson

dat <- fit[rowData(sce)$whitelist, ]

hvg <- getTopHVGs(dat, var.field = "bio", n = 2000)

sel <- ifelse(rownames(fit) %in% hvg, TRUE, FALSE)

metadata(sce)$modelGeneVarByPoisson$hvg <- sel
```

Plot expression of highly variable genes, based on modelGeneVarByPoisson statistics:

```{r}
plotHeatmap(sce, hvg, center = TRUE, symmetric = TRUE, show_rownames = FALSE, show_colnames = FALSE)
```

## Comparison

### Euler

Compare number of features selected using an Euler diagram:

```{r}
sel <- cbind(
  "Residual" = TRUE,
  "Var" = metadata(sce)$modelGeneVar$hvg,
  "CV2" = metadata(sce)$modelGeneCV2$hvg,
  "VarWithSpikes" = metadata(sce)$modelGeneVarWithSpikes$hvg,
  "CV2WithSpikes" = metadata(sce)$modelGeneCV2WithSpikes$hvg,
  "VarByPoisson" = metadata(sce)$modelGeneVarByPoisson$hvg
)

fit <- euler(sel)

pal <- RColorBrewer::brewer.pal(n = 5, name = "Pastel1")

lab <- comma(fit$original.values)

plot(fit, fills = pal, quantities = lab)
```

### Jaccard

Calculate Jaccard index:

```{r}
sel <- rbind(
  "Var" = metadata(sce)$modelGeneVar$hvg,
  "CV2" = metadata(sce)$modelGeneCV2$hvg,
  "VarWithSpikes" = metadata(sce)$modelGeneVarWithSpikes$hvg,
  "CV2WithSpikes" = metadata(sce)$modelGeneCV2WithSpikes$hvg,
  "VarByPoisson" = metadata(sce)$modelGeneVarByPoisson$hvg
)

dst <- proxy::dist(sel, method = "Jaccard", diag = TRUE, upper = TRUE)

mat <- as.matrix(dst)

pal <- RColorBrewer::brewer.pal(n = 5, name = "Greens")

col <- colorRampPalette(pal)(10)

brk <- seq(0, 1, length.out = 11)

pheatmap(mat, color = col, breaks = brk)
```

### Dimensionality reduction {.tabset}

```{r}
fit <- c("modelGeneVar", "modelGeneCV2", "modelGeneVarWithSpikes", "modelGeneCV2WithSpikes", "modelGeneVarByPoisson")
```

#### PCA

```{r, fig.height = 11, fig.width = 10}
plt <- lapply(fit, function(x) {
  
  set.seed(1701)
  
  hvg <- metadata(sce)[[x]]$hvg
  
  sce <- runPCA(sce, subset_row = hvg, BSPARAM = ExactParam())

  plotPCA(sce, colour_by = "sum") + ggtitle(x)

})

patchwork::wrap_plots(plt, ncol = 2)
```

#### TSNE

```{r, fig.height = 11, fig.width = 10}
plt <- lapply(fit, function(x) {
  
  set.seed(1701)
  
  hvg <- metadata(sce)[[x]]$hvg
  
  sce <- runPCA(sce, subset_row = hvg, BSPARAM = ExactParam())
  
  sce <- runTSNE(sce, dimred = "PCA")

  plotTSNE(sce, colour_by = "sum") + ggtitle(x)

})

patchwork::wrap_plots(plt, ncol = 2)
```

#### UMAP

```{r, fig.height = 11, fig.width = 10}
plt <- lapply(fit, function(x) {
  
  set.seed(1701)
  
  hvg <- metadata(sce)[[x]]$hvg
  
  sce <- runPCA(sce, subset_row = hvg, BSPARAM = ExactParam())
  
  sce <- runUMAP(sce, dimred = "PCA")

  plotUMAP(sce, colour_by = "sum") + ggtitle(x)

})

patchwork::wrap_plots(plt, ncol = 2)
```

## Selection

Store full dataset while we perform our downstream operations conveniently on the HVG subset:

```{r}
altExp(sce, "original") <- sce

fit <- metadata(sce)$modelGeneVarWithSpikes

hvg <- rownames(fit)[fit$hvg]

sce <- sce[hvg, ]
```

## Summary

### Output

```{r output}
write_rds(sce, here::here("data/03-feature-selection.rds"))
```

### Session

```{r session}
devtools::session_info()
```
