==> analysis/04-reduced-dimensions.Rmd <==
---
title: "Dimensionality reduction"
author: "James Ashmore"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output: workflowr::wflow_html
editor_options:
  chunk_output_type: console
---

## Motivation

In this document we aim to reduce the number of separate dimensions in the data by performing dimenstionality reduction.

## Setup

Set chunk options:

```{r knitr}
knitr::opts_chunk$set(
  autodep    = TRUE,
  cache      = TRUE,
  cache.path = "cache/04-reduced-dimensions.Rmd/",
  dev        = "png",
  error      = FALSE,
  message    = FALSE,
  warning    = FALSE
)
```

Load required packages:

```{r pacman}
pacman::p_load(
  BiocSingular,
  here,
  patchwork,
  readr,
  scater,
  scran
)
```

Read experiment data:

```{r}
rds <- here("data/03-feature-selection.rds")

sce <- read_rds(rds)
```

## PCA

### Calculate PCA

Perform a principal components analysis (PCA) on the expression data:

```{r}
set.seed(1701)

sce <- runPCA(sce, BSPARAM = ExactParam())
```

### Find elbow

Find the elbow point in the percentage of variance explained by successive PCs:

```{r}
dim <- reducedDim(sce, 'PCA')

var <- attr(dim, 'percentVar')

num <- PCAtools::findElbowPoint(var)
```

Plot the percentage of variance explained by successive PCs:

```{r}
dat <- data.frame(index = seq_along(var), total = var)

ggplot(dat, aes(index, total)) +
  geom_point() +
  geom_vline(xintercept = num, colour = 'red') +
  labs(x = 'PC', y = 'Variance (%)') +
  theme_minimal()
```

### Technical

Retain all PCs until the percentage of total variation explained reaches a given threshold based on the technical noise:

```{r}
set.seed(1701)

var <- attr(dim, 'percentVar')

mod <- metadata(sce)$modelGeneVarWithSpikes

dim <- getDenoisedPCs(sce, technical = mod, BSPARAM = ExactParam())

num <- ncol(dim$components)
```

Plot the cumulative percentage of variance explained by successive PCs:

```{r}
dat <- data.frame(index = seq_along(var), total = cumsum(var))

ggplot(dat, aes(index, total)) +
  geom_point() +
  geom_vline(xintercept = num, colour = 'red') +
  labs(x = 'PC', y = 'Cumulative variance (%)') +
  theme_minimal()
```

### Structure

Find the number of subpopulations in the data using clustering:

```{r}
dim <- reducedDim(sce, 'PCA')

len <- seq_len(ncol(dim))

res <- sapply(len, function(x) {

  mat <- dim[, seq_len(x), drop = FALSE]

  snn <- buildSNNGraph(mat, transposed = TRUE)

  cls <- igraph::cluster_walktrap(snn)$membership

  val <- length(unique(cls))

})

num <- len[which.min(abs(res - 1 - len))]
```

Plot the percentage of variance explained by successive PCs:

```{r}
dat <- data.frame(index = len, total = res)

ggplot(dat, aes(index, total)) +
  geom_point() +
  geom_vline(xintercept = num, colour = 'red') +
  labs(x = 'PC', y = 'Number of clusters') +
  theme_minimal()
```

### Selection

Choose number of PCs to retain from the above analyses:

```{r}
num <- 4

len <- seq_len(num)

dim <- reducedDim(sce, 'PCA')

reducedDim(sce, 'PCA') <- dim[, len]
```

Plot scatterplot matrix of retained PCs:

```{r, fig.height = 8, fig.width = 10}
plotPCA(sce, colour_by = "sum", ncomponents = num)
```

## TSNE

### Optimal perplexity

Inspect TSNE plots generated by different perplexity values:

```{r, fig.height = 8, fig.width = 10}
lim <- floor((ncol(sce) - 1 ) / 3)

num <- floor(seq(5, lim, length.out = 16))

plt <- lapply(num, function(x) {

  set.seed(1701)

  run <- runTSNE(sce, dimred = "PCA", perplexity = x)

  txt <- paste('Perplexity:', x)

  plotTSNE(run, colour_by = "sum") + ggtitle(txt)

})

wrap_plots(plt, ncol = 4)
```

### Optimal iterations

Inspect TSNE plots generated by different numbers of iterations:

```{r, fig.height = 8, fig.width = 10}
num <- seq(500, 5000, length.out = 16)

plt <- lapply(num, function(x) {

  set.seed(1701)

  run <- runTSNE(sce, dimred = 'PCA', max_iter = x)

  txt <- paste('Iterations:', x)

  plotTSNE(run, colour_by = 'sum') + ggtitle(txt)

})

wrap_plots(plt, ncol = 4)
```

### Calculate TSNE

Select optimal perplexity value and number of iterations:

```{r}
set.seed(1701)

sce <- runTSNE(sce, dimred = "PCA", perplexity = 23, max_iter = 1000)
```

## UMAP

Perform uniform manifold approximation and projection (UMAP) on the PCA data.

### Optimal neighbors

```{r, fig.height = 8, fig.width = 10}
num <- seq(5, 50, length.out = 16)

plt <- lapply(num, function(n) {

  set.seed(1701)

  run <- runUMAP(sce, dimred = 'PCA', n_neighbors = n)

  txt <- paste('Neighbors:', n)

  plotUMAP(run, colour_by = 'sum') + ggtitle(txt)

})

wrap_plots(plt, ncol = 4)
```

### Calculate UMAP

Select optimal number of nearest neighbors:

```{r}
set.seed(1701)

sce <- runUMAP(sce, dimred = 'PCA', n_neighbors = 29)
```

## Variables {.tabset}

### PCA

```{r, fig.height = 8, fig.width = 10}
var <- c("sum", "detected", "subsets_MT_percent", "altexps_ERCC_percent")

plt <- lapply(var, function(x) plotPCA(sce, colour_by = x))

patchwork::wrap_plots(plt, ncol = 2)
```

### TSNE

```{r, fig.height = 8, fig.width = 10}
var <- c("sum", "detected", "subsets_MT_percent", "altexps_ERCC_percent")

plt <- lapply(var, function(x) plotTSNE(sce, colour_by = x))

patchwork::wrap_plots(plt, ncol = 2)
```

### UMAP

```{r, fig.height = 8, fig.width = 10}
var <- c("sum", "detected", "subsets_MT_percent", "altexps_ERCC_percent")

plt <- lapply(var, function(x) plotUMAP(sce, colour_by = x))

patchwork::wrap_plots(plt, ncol = 2)
```

## Summary

### Output

Write experiment data:

```{r output}
write_rds(sce, here("data/04-reduced-dimensions.rds"))
```

### Session

Print session information:

```{r session}
devtools::session_info()
```
