==> analysis/09-cell-annotation.Rmd <==
---
title: "Cell type annotation"
author: "James Ashmore"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output: workflowr::wflow_html
editor_options:
  chunk_output_type: console
---

## Introduction

In this analysis we are going to annotate previously defined clusters with a cell type label. This process can be done manually using a domain expert's knowledge or automatically based on known markers and reference data.

## Setup

Define chunk options:

```{r knitr}
knitr::opts_chunk$set(
  autodep    = TRUE,
  cache      = TRUE,
  cache.path = "cache/09-cell-annotation.Rmd/",
  cache.lazy = FALSE,
  dev        = "png",
  error      = FALSE,
  message    = FALSE,
  warning    = FALSE,
  fig.align = "center",
  out.width = '100%'
)
```

Attach required packages:

```{r pacman}
pacman::p_load(
  AUCell,
  SingleR,
  GSEABase,
  scran,
  scater,
  here,
  readr,
  patchwork,
  pheatmap
)
```

Import experiment data:

```{r}
rds <- here("data/08-marker-detection.rds")

sce <- read_rds(rds)

alt <- altExp(sce, "original", withColData = TRUE)
```

## Annotation

### Known markers

Rename features of test data for compatability with known markers:

```{r}
rownames(alt) <- rowData(alt)$gene_name
```

Read marker sets from disk and check they are present in the data:

```{r}
all <- read_csv("data/markers.csv")

all <- all[all$gene_name %in% rownames(alt), ]
```

Format marker sets for entry into the AUCell package:

```{r}
all <- split(all, all$gene_set)

all <- lapply(all, function(x) GeneSet(x$gene_name, setName = unique(x$gene_set)))

all <- GeneSetCollection(all)
```

Use the AUCell package to identify markers sets that are highly expressed in each cell:

```{r}
fit <- AUCell_buildRankings(counts(alt), plotStats = FALSE, verbose = FALSE)

auc <- AUCell_calcAUC(all, fit)
```

Get assignments from AUCell scores:

```{r}
res <- t(assay(auc))

lab <- colnames(res)[max.col(res)]
```

Store AUCell assignments in experiment data:

```{r}
sce$celltype_aucell <- lab
```

Create results table for use with SingleR functions:

```{r}
dat <- DataFrame(scores = I(res), labels = lab)

dat <- dat[rowSds(dat$scores) > 0, ]
```

Create a heatmap of the AUCell assignment scores:

```{r}
plotScoreHeatmap(dat, clusters = sce$cluster, fontsize = 8)
```

Plot score distributions of labels:

```{r}
plotScoreDistribution(dat, show = "delta.med", ncol = 3, show.nmads = 3)
```

### Gastrulation data

Rename features of test data for compatability with reference data:

```{r}
rownames(alt) <- rowData(alt)$gene_id
```

Load the gastrulation data package:

```{r}
library(MouseGastrulationData)
```

Process E8.25 samples from gastrulation data:

```{r}
sel <- subset(AtlasSampleMetadata, stage == "E8.25")

ref <- EmbryoAtlasData(type = "processed", samples = sel$sample)

ref <- logNormCounts(ref)
```

Remove cells with NA values for the given celltype label:

```{r}
sel <- !is.na(ref$celltype)

ref <- ref[, sel]
```

Remove cells marked as stripped nuclei:

```{r}
sel <- !is.na(ref$stripped)

ref <- ref[, sel]
```

Remove cells marked as doublets:

```{r}
sel <- !is.na(ref$doublet)

ref <- ref[, sel]
```

Identify marker genes for each cell type whilst blocking on sample:

```{r}
res <- pairwiseWilcox(logcounts(ref), groups = ref$celltype, block = ref$sample)

sig <- getTopMarkers(res$statistics, res$pairs)
```

Aggregate references within each celltype and sample combination:

```{r}
lab <- paste(ref$celltype, ref$sample, sep = "-")

ref <- aggregateReference(ref, labels = lab)
```

Remove sample label from aggregated reference names:

```{r}
lab <- sapply(strsplit(colnames(ref), "-"), '[', 1)
```

Return the best annotation for each cell in the test dataset:

```{r}
fit <- SingleR(alt, ref, labels = lab, genes = sig)
```

Store SingleR assignments in experiment data:

```{r}
sce$celltype_singler <- fit$labels
```

Create a heatmap of the average SingleR assignment scores:

```{r}
plotScoreHeatmap(fit, clusters = alt$cluster, fontsize = 8)
```

Plot score distributions of labels:

```{r fig.height = 20}
plotScoreDistribution(fit, show = "delta.med", ncol = 3, show.nmads = 3)
```

## Selection

Choose cell type annotation method for downstream analysis:

```{r}
sce$celltype <- sce$celltype_singler
```

## Summary

### Output

Write experiment data:

```{r output}
write_rds(sce, here("data/09-cell-annotation.rds"))
```

### Session

Print session information:

```{r session}
devtools::session_info()
```
